+++ core/src/gaia/cu9/ari/gaiaorbit/data/stars/OctreeCatalogLoader.java
@@ -70,7 +70,7 @@ public class OctreeCatalogLoader implements ISceneGraphLoader {
          * MANUALLY ADD SUN
          */
         // Manually add sun
-        Star sun = new Star(new Vector3d(0, 0, 0), 4.83f, 4.83f, 0.656f, "Sol", System.currentTimeMillis());
+        Star sun = new Star(new Vector3d(0, 0, 0), 4.83f, 4.83f, 0.656f, "Sol", (int) System.currentTimeMillis());
         sun.initialize();
 
         // Find out octant of sun
